+++ django_project/geohosting/src/pages/CheckoutPage/CheckoutPage/index.tsx
@@ -43,9 +43,10 @@ interface CheckoutPageModalProps {
   paystackUrl: string;
 }
 
-const MainCheckoutPage: React.FC<CheckoutPageModalProps> = (
+export const MainCheckoutPageComponent: React.FC<CheckoutPageModalProps> = (
   { product, pkg, stripeUrl, paystackUrl }
 ) => {
+  /** For the payment component **/
   const columns = useBreakpointValue({ base: 1, md: 2 });
   const [paymentMethod, setPaymentMethod] = useState<string>(PaymentMethods.STRIPE);
   const stripePaymentModalRef = useRef(null);
@@ -72,6 +73,82 @@ const MainCheckoutPage: React.FC<CheckoutPageModalProps> = (
   }
 
   return (
+    <>
+      <Grid gap={6} templateColumns={`repeat(${columns}, 1fr)`}>
+        <GridItem>
+          <Box>
+            <Text fontSize={22} color={'black'}>
+              Payment Method
+            </Text>
+          </Box>
+          <Box padding={8} backgroundColor="gray.100" borderRadius={10}>
+            <VStack spacing={4} align="stretch">
+              <Box border="1px" borderColor="gray.300" borderRadius="md"
+                   p="4">
+                <HStack justifyContent="space-between">
+                  <HStack>
+                    <FaCcStripe size="30px"/>
+                  </HStack>
+                  <IconButton aria-label={'icon'} variant="ghost"/>
+                </HStack>
+                <Text mt={2}>
+                  By purchasing this subscription and clicking
+                  "Continue", you agree to the <Link href="#">terms of
+                  service</Link>, <Link href="#">auto-renewal
+                  terms</Link>, electronic document delivery, and
+                  acknowledge the <Link href="#">privacy policy</Link>.
+                </Text>
+                <Button
+                  mt={4} leftIcon={<FaCcStripe/>} mr={1}
+                  colorScheme={paymentMethod === PaymentMethods.STRIPE ? "blue" : "blackAlpha"}
+                  size="lg"
+                  onClick={() => setPaymentMethod(PaymentMethods.STRIPE)}
+                >
+                  Pay with Stripe
+                </Button>
+                <Button
+                  mt={4}
+                  colorScheme={paymentMethod === PaymentMethods.PAYSTACK ? "blue" : "blackAlpha"}
+                  size="lg"
+                  onClick={() => setPaymentMethod(PaymentMethods.PAYSTACK)}
+                >
+                  Pay with Paystack
+                </Button>
+                <Divider mt={4}/>
+                <Text mt={2} fontSize="sm">Payments are processed
+                  in {pkg.currency}. Payment provider fees may
+                  apply.</Text>
+              </Box>
+            </VStack>
+          </Box>
+        </GridItem>
+        <OrderSummary product={product} pkg={pkg}/>
+      </Grid>
+      <Box mt={4}>
+        <Button
+          w='100%'
+          colorScheme="orange"
+          onClick={checkout}
+        >
+          Pay with {paymentMethod.toLowerCase()}
+        </Button>
+      </Box>
+      <StripePaymentModal
+        ref={stripePaymentModalRef}
+        url={stripeUrl}
+      />
+      <PaystackPaymentModal
+        ref={paystackPaymentModalRef}
+        url={paystackUrl}
+      />
+    </>
+  )
+}
+
+const MainCheckoutPage: React.FC<CheckoutPageModalProps> = (
+  { product, pkg, stripeUrl, paystackUrl }
+) => {
+  return (
     <ChakraProvider theme={customTheme}>
       <Flex direction="column" minHeight="100vh">
         <Box flex="1">
@@ -81,65 +158,10 @@ const MainCheckoutPage: React.FC<CheckoutPageModalProps> = (
             <Box mb={10}>
               <CheckoutTracker activeStep={0}/>
             </Box>
-            <Grid gap={6} templateColumns={`repeat(${columns}, 1fr)`}>
-              <GridItem>
-                <Box>
-                  <Text fontSize={22} color={'black'}>
-                    Payment Method
-                  </Text>
-                </Box>
-                <Box padding={8} backgroundColor="gray.100" borderRadius={10}>
-                  <VStack spacing={4} align="stretch">
-                    <Box border="1px" borderColor="gray.300" borderRadius="md"
-                         p="4">
-                      <HStack justifyContent="space-between">
-                        <HStack>
-                          <FaCcStripe size="30px"/>
-                        </HStack>
-                        <IconButton aria-label={'icon'} variant="ghost"/>
-                      </HStack>
-                      <Text mt={2}>
-                        By purchasing this subscription and clicking
-                        "Continue", you agree to the <Link href="#">terms of
-                        service</Link>, <Link href="#">auto-renewal
-                        terms</Link>, electronic document delivery, and
-                        acknowledge the <Link href="#">privacy policy</Link>.
-                      </Text>
-                      <Button
-                        mt={4} leftIcon={<FaCcStripe/>} mr={1}
-                        colorScheme={paymentMethod === PaymentMethods.STRIPE ? "blue" : "blackAlpha"}
-                        size="lg"
-                        onClick={() => setPaymentMethod(PaymentMethods.STRIPE)}
-                      >
-                        Pay with Stripe
-                      </Button>
-                      <Button
-                        mt={4}
-                        colorScheme={paymentMethod === PaymentMethods.PAYSTACK ? "blue" : "blackAlpha"}
-                        size="lg"
-                        onClick={() => setPaymentMethod(PaymentMethods.PAYSTACK)}
-                      >
-                        Pay with Paystack
-                      </Button>
-                      <Divider mt={4}/>
-                      <Text mt={2} fontSize="sm">Payments are processed
-                        in {pkg.currency}. Payment provider fees may
-                        apply.</Text>
-                    </Box>
-                  </VStack>
-                </Box>
-              </GridItem>
-              <OrderSummary product={product} pkg={pkg}/>
-            </Grid>
-            <Box mt={4}>
-              <Button
-                w='100%'
-                colorScheme="orange"
-                onClick={checkout}
-              >
-                Pay with {paymentMethod.toLowerCase()}
-              </Button>
-            </Box>
+            <MainCheckoutPageComponent
+              product={product} pkg={pkg}
+              stripeUrl={stripeUrl}
+              paystackUrl={paystackUrl}/>
           </Container>
         </Box>
         <Box
@@ -150,14 +172,6 @@ const MainCheckoutPage: React.FC<CheckoutPageModalProps> = (
         >
           <Text color="white">Powered by Kartoza</Text>
         </Box>
-        <StripePaymentModal
-          ref={stripePaymentModalRef}
-          url={stripeUrl}
-        />
-        <PaystackPaymentModal
-          ref={paystackPaymentModalRef}
-          url={paystackUrl}
-        />
       </Flex>
     </ChakraProvider>
   );
